stash: port benchmarks to async api
